feat: update Java models with dynamic env. variables, refactoring
