+++ api-model/.gitignore
@@ -48,6 +48,4 @@ target/
 website
 
 # Other
-*~
-.notes
 **/bin
