+++ api-model/pom.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-    xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
     <groupId>io.apicurio</groupId>
@@ -69,13 +69,15 @@
 
         <version.fabric8>5.12.1</version.fabric8>
         <sundrio.version>0.40.1</sundrio.version>
+        <version.slf4j>1.7.36</version.slf4j>
 
         <!-- Plugins -->
         <version.maven-javadoc-plugin>3.3.1</version.maven-javadoc-plugin>
         <version.maven-source-plugin>3.2.1</version.maven-source-plugin>
         <version.maven-compiler-plugin>3.9.0</version.maven-compiler-plugin>
         <surefire-plugin.version>2.22.2</surefire-plugin.version>
-        <surefire-junit.version>1.3.2</surefire-junit.version>
+        <surefire-junit.version>5.9.0</surefire-junit.version>
+        <junit.platform-launcher.version>1.9.0</junit.platform-launcher.version>
     </properties>
 
     <dependencies>
@@ -113,7 +115,35 @@
         <dependency>
             <groupId>org.projectlombok</groupId>
             <artifactId>lombok</artifactId>
-            <version>1.18.22</version>
+            <version>1.18.24</version>
+        </dependency>
+        <dependency>
+            <groupId>org.slf4j</groupId>
+            <artifactId>slf4j-api</artifactId>
+            <version>${version.slf4j}</version>
+        </dependency>
+        <dependency>
+            <groupId>org.slf4j</groupId>
+            <artifactId>slf4j-simple</artifactId>
+            <version>${version.slf4j}</version>
+        </dependency>
+        <dependency>
+            <groupId>org.junit.jupiter</groupId>
+            <artifactId>junit-jupiter</artifactId>
+            <version>${surefire-junit.version}</version>
+            <scope>test</scope>
+        </dependency>
+        <dependency>
+            <groupId>org.junit.jupiter</groupId>
+            <artifactId>junit-jupiter-api</artifactId>
+            <version>${surefire-junit.version}</version>
+            <scope>provided</scope>
+        </dependency>
+        <dependency>
+            <groupId>org.junit.platform</groupId>
+            <artifactId>junit-platform-launcher</artifactId>
+            <version>${junit.platform-launcher.version}</version>
+            <scope>provided</scope>
         </dependency>
     </dependencies>
 
@@ -160,9 +190,14 @@
                     <source>${maven.compiler.source}</source>
                 </configuration>
             </plugin>
+            <plugin>
+                <groupId>org.apache.maven.plugins</groupId>
+                <artifactId>maven-surefire-plugin</artifactId>
+                <version>${surefire-plugin.version}</version>
+            </plugin>
         </plugins>
     </build>
-    
+
     <profiles>
         <profile>
             <id>release</id>
@@ -209,5 +244,5 @@
             </build>
         </profile>
     </profiles>
-    
+
 </project>
