--- api-model/src/main/java/io/apicurio/registry/operator/api/model/ApicurioRegistrySpecConfigurationKafkaSecurityScram.java
@@ -1,64 +1,0 @@
-/*
- * Copyright 2022 Red Hat
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package io.apicurio.registry.operator.api.model;
-
-import io.sundr.builder.annotations.Buildable;
-import lombok.EqualsAndHashCode;
-
-@Buildable(
-        editableEnabled = false,
-        builderPackage = Constants.FABRIC8_KUBERNETES_API
-)
-@EqualsAndHashCode
-public class ApicurioRegistrySpecConfigurationKafkaSecurityScram {
-    String truststoreSecretName;
-    String user;
-    String passwordSecretName;
-    String mechanism;
-
-    public String getTruststoreSecretName() {
-        return truststoreSecretName;
-    }
-
-    public void setTruststoreSecretName(String truststoreSecretName) {
-        this.truststoreSecretName = truststoreSecretName;
-    }
-
-    public String getUser() {
-        return user;
-    }
-
-    public void setUser(String user) {
-        this.user = user;
-    }
-
-    public String getPasswordSecretName() {
-        return passwordSecretName;
-    }
-
-    public void setPasswordSecretName(String passwordSecretName) {
-        this.passwordSecretName = passwordSecretName;
-    }
-
-    public String getMechanism() {
-        return mechanism;
-    }
-
-    public void setMechanism(String mechanism) {
-        this.mechanism = mechanism;
-    }
-}
