+++ api-model/src/main/java/io/apicurio/registry/operator/api/v1/model/ApicurioRegistrySpecConfigurationKafkaSecurityTls.java
@@ -14,24 +14,38 @@
  * limitations under the License.
  */
 
-package io.apicurio.registry.operator.api.model;
+package io.apicurio.registry.operator.api.v1.model;
 
 import io.sundr.builder.annotations.Buildable;
 import lombok.EqualsAndHashCode;
+import lombok.ToString;
 
 @Buildable(
         editableEnabled = false,
         builderPackage = Constants.FABRIC8_KUBERNETES_API
 )
 @EqualsAndHashCode
-public class ApicurioRegistrySpecConfigurationSecurity {
-    private ApicurioRegistrySpecConfigurationSecurityKeycloak keycloak;
+@ToString
+// NOTE: We can not use Lombok @Getter and @Setter because it does not work with fabric8 generator.
+public class ApicurioRegistrySpecConfigurationKafkaSecurityTls {
 
-    public ApicurioRegistrySpecConfigurationSecurityKeycloak getKeycloak() {
-        return keycloak;
+    private String truststoreSecretName;
+
+    private String keystoreSecretName;
+
+    public String getTruststoreSecretName() {
+        return truststoreSecretName;
     }
 
-    public void setKeycloak(ApicurioRegistrySpecConfigurationSecurityKeycloak keycloak) {
-        this.keycloak = keycloak;
+    public void setTruststoreSecretName(String truststoreSecretName) {
+        this.truststoreSecretName = truststoreSecretName;
+    }
+
+    public String getKeystoreSecretName() {
+        return keystoreSecretName;
+    }
+
+    public void setKeystoreSecretName(String keystoreSecretName) {
+        this.keystoreSecretName = keystoreSecretName;
     }
 }
