+++ api-model/src/main/java/io/apicurio/registry/operator/api/v1/model/ApicurioRegistrySpecConfigurationSql.java
@@ -14,35 +14,28 @@
  * limitations under the License.
  */
 
-package io.apicurio.registry.operator.api.model;
-
-import java.util.HashMap;
+package io.apicurio.registry.operator.api.v1.model;
 
 import io.sundr.builder.annotations.Buildable;
 import lombok.EqualsAndHashCode;
+import lombok.ToString;
 
 @Buildable(
         editableEnabled = false,
         builderPackage = Constants.FABRIC8_KUBERNETES_API
 )
 @EqualsAndHashCode
-public class ApicurioRegistrySpecDeploymentMetadata {
-    HashMap<String, String> annotations;
-    HashMap<String, String> labels;
+@ToString
+// NOTE: We can not use Lombok @Getter and @Setter because it does not work with fabric8 generator.
+public class ApicurioRegistrySpecConfigurationSql {
 
-    public HashMap<String, String> getAnnotations() {
-        return annotations;
+    ApicurioRegistrySpecConfigurationDataSource dataSource;
+
+    public ApicurioRegistrySpecConfigurationDataSource getDataSource() {
+        return dataSource;
     }
 
-    public void setAnnotations(HashMap<String, String> annotations) {
-        this.annotations = annotations;
-    }
-
-    public HashMap<String, String> getLabels() {
-        return labels;
-    }
-
-    public void setLabels(HashMap<String, String> labels) {
-        this.labels = labels;
+    public void setDataSource(ApicurioRegistrySpecConfigurationDataSource dataSource) {
+        this.dataSource = dataSource;
     }
 }
