+++ api-model/src/test/resources/apicurio-registry-cr.yaml
@@ -1,0 +1,11 @@
+apiVersion: registry.apicur.io/v1
+kind: ApicurioRegistry
+metadata:
+  name: test
+  namespace: test-namespace
+spec:
+  configuration:
+    persistence: mem
+    env:
+      - name: VAR_1_NAME
+        value: VAR_1_VALUE
